Moved Code to Puzzle Manager
Moved code for piece manipulation to operate inside of puzzle manager instead of input manager.
Input Manager now calls functions from its local copy of the Puzzle Manager.

All tested/working
